+++ src/data_types/compare_benchmark.rs
@@ -1,9 +1,6 @@
 
-use anyhow::{bail, ensure};
-use std::collections::BTreeMap;
-
 use crate::data_types::grouped_metrics::GroupMetrics;
-use crate::data_types::summary_metrics::{SummaryGtMetrics, SummaryMetrics};
+use crate::data_types::summary_metrics::SummaryMetrics;
 use crate::data_types::variant_metrics::{VariantMetrics, VariantSource};
 use crate::data_types::variants::{Variant, VariantType};
 
@@ -16,19 +13,9 @@ pub struct CompareBenchmark {
     bm_edit_distance_h1: usize,
     /// Best match, edit distance for hap2
     bm_edit_distance_h2: usize,
-    /// Best match genotype metrics
-    bm_gt: SummaryGtMetrics,
-    /// Best match haplotype metrics
-    bm_hap: SummaryMetrics,
-    /// Best match basepair metrics
-    bm_basepair: SummaryMetrics,
 
-    /// Variant genotype statistics
-    variant_gt: BTreeMap<VariantType, SummaryGtMetrics>,
-    /// Variant haplotype statistics
-    variant_hap: BTreeMap<VariantType, SummaryMetrics>,
-    /// Variant basepair statistics
-    variant_basepair: BTreeMap<VariantType, SummaryMetrics>,
+    /// Grouped metrics, which are the bulk of tracked statistics
+    group_metrics: GroupMetrics,
 
     // variant level statistics
     /// Variant metrics for truth variants
@@ -58,12 +45,7 @@ impl CompareBenchmark {
             region_id,
             bm_edit_distance_h1,
             bm_edit_distance_h2,
-            bm_gt: Default::default(),
-            bm_hap: Default::default(),
-            bm_basepair: Default::default(),
-            variant_gt: Default::default(),
-            variant_hap: Default::default(),
-            variant_basepair: Default::default(),
+            group_metrics: Default::default(),
             truth_variant_data: Default::default(),
             query_variant_data: Default::default(),
             sequence_bundle: None,
@@ -77,51 +59,8 @@ impl CompareBenchmark {
     /// * `expected_zygosity_count` - number of times this allele was expected in the truth set; e.g. 0/1 => 1, 1/1 => 2
     /// * `observed_zygosity_count` - number of times this allele was identified in the best-matching query haplotype sequences
     pub fn add_truth_zygosity(&mut self, variant: &Variant, expected_zygosity_count: u8, observed_zygosity_count: u8) -> anyhow::Result<()> {
-        // we are assuming that we're only looking at non-reference genotypes
-        assert!(expected_zygosity_count > 0);
-
-        // get the variant type for classification, and the variant metrics for mutation
-        let variant_type = variant.variant_type();
-        let v_gt = self.variant_gt.entry(variant_type).or_default();
-        let v_hap = self.variant_hap.entry(variant_type).or_default();
-
-        /*
-        In our modified exact setup, variants are either found or they are not; meaning that a false positive does not exist from a "truth" perspective.
-        Instead, a false positive is just a query variant that is not found in truth (i.e., the inverse calculations).
-        This means we should bail! if we find more observed alleles than expected.
-         */
-
-        // compare expected to observed
-        match expected_zygosity_count.cmp(&observed_zygosity_count) {
-            std::cmp::Ordering::Less => {
-                // we found too many observed relative to expected, so add the extra to FP
-                bail!("No implementation for truth false positives");
-            },
-            std::cmp::Ordering::Equal => {
-                // they match, so add one per expected
-                self.bm_hap.truth_tp += expected_zygosity_count as u64;
-                v_hap.truth_tp += expected_zygosity_count as u64;
-
-                self.bm_gt.summary_metrics.truth_tp += 1;
-                v_gt.summary_metrics.truth_tp += 1;
-            },
-            std::cmp::Ordering::Greater => {
-                // we found too few relative to expected, so add the missing to FN
-                self.bm_hap.truth_tp += observed_zygosity_count as u64;
-                self.bm_hap.truth_fn += (expected_zygosity_count - observed_zygosity_count) as u64;
-                v_hap.truth_tp += observed_zygosity_count as u64;
-                v_hap.truth_fn += (expected_zygosity_count - observed_zygosity_count) as u64;
-
-                // expected > observed, this is a false negative
-                self.bm_gt.summary_metrics.truth_fn += 1;
-                v_gt.summary_metrics.truth_fn += 1;
-                if observed_zygosity_count > 0 {
-                    // observed was >0, this is purely a GT difference like 1/1 -> 0/1
-                    self.bm_gt.truth_fn_gt += 1;
-                    v_gt.truth_fn_gt += 1;
-                }
-            },
-        };
+        // add to the group metrics
+        self.group_metrics.add_truth_zygosity(variant, expected_zygosity_count, observed_zygosity_count)?;
 
         // add the variant metrics
         let variant_metrics = VariantMetrics::new(VariantSource::Truth, expected_zygosity_count, observed_zygosity_count)?;
@@ -136,23 +75,8 @@ impl CompareBenchmark {
     /// * `expected_zygosity_count` - number of times this allele was expected in the query set; e.g. 0/1 => 1, 1/1 => 2
     /// * `observed_zygosity_count` - number of times this allele was identified in the best-matching truth haplotype sequences
     pub fn add_query_zygosity(&mut self, variant: &Variant, expected_zygosity_count: u8, observed_zygosity_count: u8) -> anyhow::Result<()> {
-        // we are assuming that we're only looking at non-reference genotypes
-        assert!(expected_zygosity_count > 0);
-
-        // get the variant type for classification, and the variant metrics for mutation
-        let variant_type = variant.variant_type();
-        let v_gt = self.variant_gt.entry(variant_type).or_default();
-        let v_hap = self.variant_hap.entry(variant_type).or_default();
-
-        // normally, we would compare expected to observed, but this function is only used for exact matches currently
-        ensure!(expected_zygosity_count == observed_zygosity_count, "No implementation for non-equal query zygosities");
-
-        // they match, so add one per expected
-        self.bm_hap.query_tp += expected_zygosity_count as u64;
-        v_hap.query_tp += expected_zygosity_count as u64;
-
-        self.bm_gt.summary_metrics.query_tp += 1;
-        v_gt.summary_metrics.query_tp += 1;
+        // add to the group stats
+        self.group_metrics.add_query_zygosity(variant, expected_zygosity_count, observed_zygosity_count)?;
 
         // add the variant metrics, since this is a query variant, toggle the source info
         let variant_metrics = VariantMetrics::toggle_source(
@@ -168,35 +92,15 @@ impl CompareBenchmark {
     /// * `basepair_metrics` - the metrics to get added into the current values
     /// * `variant_type` - optional variant type these stats are added to; if none, it goes into the "All" grouping
     pub fn add_basepair_metrics(&mut self, basepair_metrics: SummaryMetrics, variant_type: Option<VariantType>) {
-        if let Some(vt) = variant_type {
-            let v_basepair = self.variant_basepair.entry(vt).or_default();
-            *v_basepair += basepair_metrics;
-        } else {
-            // all variant category
-            self.bm_basepair += basepair_metrics;
-        }
+        self.group_metrics.add_basepair_metrics(basepair_metrics, variant_type);
     }
 
     /// Sets the values for a reverse benchmark. This is primarily to set query_tp and query_fp values from a reverse comparison.
     /// # Arguments
     /// * `other` - Results from a benchmark where truth and query have been swapped.
     pub fn add_swap_benchmark(&mut self, other: &Self) -> anyhow::Result<()> {
-        // set the best match values
-        self.bm_gt.set_query_from_truth(&other.bm_gt);
-        self.bm_hap.set_query_from_truth(&other.bm_hap);
-
-        // bm_basepair and variant_basepair do not compute the inverse separately
-
-        // set the per-variant values as well for both
-        for (vt, metrics) in other.variant_gt.iter() {
-            let entry = self.variant_gt.entry(*vt).or_default();
-            entry.set_query_from_truth(metrics);
-        }
-
-        for (vt, metrics) in other.variant_hap.iter() {
-            let entry = self.variant_hap.entry(*vt).or_default();
-            entry.set_query_from_truth(metrics);
-        }
+        // add the swapped group metrics first
+        self.group_metrics.add_swap_benchmark(&other.group_metrics)?;
 
         // we need to add any truth variants from other to our query variants, and also invert them as we go
         self.query_variant_data.extend(
@@ -227,29 +131,13 @@ impl CompareBenchmark {
         self.bm_edit_distance_h1 + self.bm_edit_distance_h2
     }
 
-    /// Utility function that creates the grouped metrics for us by copying out of this benchmark.
-    pub fn group_metrics(&self) -> GroupMetrics {
-        GroupMetrics::new(
-            self.bm_gt, self.bm_hap, self.bm_basepair,
-            self.variant_gt.clone(), self.variant_hap.clone(), self.variant_basepair.clone()
-        )
-    }
-
     // getters
     pub fn region_id(&self) -> u64 {
         self.region_id
     }
 
-    pub fn bm_hap(&self) -> SummaryMetrics {
-        self.bm_hap
-    }
-
-    pub fn bm_gt(&self) -> SummaryGtMetrics {
-        self.bm_gt
-    }
-
-    pub fn bm_basepair(&self) -> SummaryMetrics {
-        self.bm_basepair
+    pub fn group_metrics(&self) -> &GroupMetrics {
+        &self.group_metrics
     }
 
     pub fn truth_variant_data(&self) -> &[VariantMetrics] {
@@ -258,18 +146,6 @@ impl CompareBenchmark {
 
     pub fn query_variant_data(&self) -> &[VariantMetrics] {
         &self.query_variant_data
-    }
-
-    pub fn variant_gt(&self) -> &BTreeMap<VariantType, SummaryGtMetrics> {
-        &self.variant_gt
-    }
-
-    pub fn variant_hap(&self) -> &BTreeMap<VariantType, SummaryMetrics> {
-        &self.variant_hap
-    }
-
-    pub fn variant_basepair(&self) -> &BTreeMap<VariantType, SummaryMetrics> {
-        &self.variant_basepair
     }
 
     pub fn sequence_bundle(&self) -> Option<&SequenceBundle> {